+++ lib/shopify_app/configuration.rb
@@ -37,9 +37,6 @@ module ShopifyApp
     # allow namespacing webhook jobs
     attr_accessor :webhook_jobs_namespace
 
-    # allow enabling of same site none on cookies
-    attr_writer :enable_same_site_none
-
     def initialize
       @root_url = '/'
       @myshopify_domain = 'myshopify.com'
@@ -84,10 +81,6 @@ module ShopifyApp
 
     def has_scripttags?
       scripttags.present?
-    end
-
-    def enable_same_site_none
-      !Rails.env.test? && (@enable_same_site_none.nil? ? embedded_app? : @enable_same_site_none)
     end
 
     def shop_access_scopes