+++ lib/shopify_app/engine.rb
@@ -26,10 +26,8 @@ module ShopifyApp
     end
 
     initializer "shopify_app.middleware" do |app|
-      app.config.middleware.insert_after(::Rack::Runtime, ShopifyApp::SameSiteCookieMiddleware)
-
-      if ShopifyApp.configuration.embedded_app?
-        app.config.middleware.insert_after(ShopifyApp::SameSiteCookieMiddleware, ShopifyApp::JWTMiddleware)
+      if ShopifyApp.configuration.allow_jwt_authentication
+        app.config.middleware.insert_after(::Rack::Runtime, ShopifyApp::JWTMiddleware)
       end
     end